requirements.txt
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.2
future==0.18.2
numpy==1.18.1
opencv-python==4.2.0.32
pycparser==2.20
python-telegram-bot==12.4.2
six==1.14.0
tornado==6.0.4
face-recognition==1.3.0
pymongo==3.10.1
